requirements.txt
streamlit>=1.32.0
groq>=0.9.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
plotly>=5.18.0
